+++ src/option.ts
@@ -5,25 +5,30 @@ export type BuilderOption = {
   addAsBeforeColumnAlias?: 'always'|'never'
   addPublicSchemaName?: 'always'|'never'|'when other schema mentioned'
   addTableName?: 'always'|'when two tables or more' //TODO: add more options like 'when needed' and 'when conflict'
+  addAsBeforeTableAlias?: 'always'|'never'
 }
 
-const defaultOption: BuilderOption = {
+export type BuilderOptionRequired = Required<BuilderOption>
+
+const defaultOption: BuilderOptionRequired = {
   useSemicolonAtTheEnd: true,
   addAscAfterOrderByItem: 'when mentioned',
   addNullsLastAfterOrderByItem: 'when mentioned',
   addAsBeforeColumnAlias: 'always',
   addPublicSchemaName: 'never',
   addTableName: 'when two tables or more',
+  addAsBeforeTableAlias: 'always',
 }
 Object.freeze(defaultOption)
 
-export function fillUndefinedOptionsWithDefault(option: BuilderOption): BuilderOption {
-  const result: BuilderOption = {}
-  result.useSemicolonAtTheEnd = option.useSemicolonAtTheEnd ?? defaultOption.useSemicolonAtTheEnd
-  result.addAscAfterOrderByItem = option.addAscAfterOrderByItem ?? defaultOption.addAscAfterOrderByItem
-  result.addNullsLastAfterOrderByItem = option.addNullsLastAfterOrderByItem ?? defaultOption.addNullsLastAfterOrderByItem
-  result.addAsBeforeColumnAlias = option.addAsBeforeColumnAlias ?? defaultOption.addAsBeforeColumnAlias
-  result.addPublicSchemaName = option.addPublicSchemaName ?? defaultOption.addPublicSchemaName
-  result.addTableName = option.addTableName ?? defaultOption.addTableName
-  return result
+export function fillUndefinedOptionsWithDefault(option: BuilderOption): BuilderOptionRequired {
+  return {
+    useSemicolonAtTheEnd: option.useSemicolonAtTheEnd ?? defaultOption.useSemicolonAtTheEnd,
+    addAscAfterOrderByItem: option.addAscAfterOrderByItem ?? defaultOption.addAscAfterOrderByItem,
+    addNullsLastAfterOrderByItem: option.addNullsLastAfterOrderByItem ?? defaultOption.addNullsLastAfterOrderByItem,
+    addAsBeforeColumnAlias: option.addAsBeforeColumnAlias ?? defaultOption.addAsBeforeColumnAlias,
+    addPublicSchemaName: option.addPublicSchemaName ?? defaultOption.addPublicSchemaName,
+    addTableName: option.addTableName ?? defaultOption.addTableName,
+    addAsBeforeTableAlias: option.addAsBeforeTableAlias ?? defaultOption.addAsBeforeTableAlias,
+  }
 }