+++ test/distinctStep.test.ts
@@ -5,9 +5,9 @@ import {
 } from '../src'
 import { database } from './database'
 //Alias
-const table = database.s.public.t.testTable
-const col1 = database.s.public.t.testTable.c.col1
-const col2 = database.s.public.t.testTable.c.col2
+const table = database.s.public.t.table1
+const col1 = database.s.public.t.table1.c.col1
+const col2 = database.s.public.t.table1.c.col2
 
 
 describe('test orderBy Step', () => {
@@ -17,66 +17,66 @@ describe('test orderBy Step', () => {
   /* In Postgres it is allowed to have FROM directly
    after SELECT with or without ALL
    */
-  it('Produces [SELECT ALL FROM "testTable";]', () => {
+  it('Produces [SELECT ALL FROM "table1";]', () => {
     const actual = sql
       .selectAll()
       .from(table)
       .getSQL()
 
-    expect(actual).toEqual('SELECT ALL FROM "testTable";')
+    expect(actual).toEqual('SELECT ALL FROM "table1";')
   })
 
-  it('Produces [SELECT DISTINCT "col1", "col2" FROM "testTable";]', () => {
+  it('Produces [SELECT DISTINCT "col1", "col2" FROM "table1";]', () => {
     const actual = sql
       .selectDistinct(col1, col2)
       .from(table)
       .getSQL()
 
-    expect(actual).toEqual('SELECT DISTINCT "col1", "col2" FROM "testTable";')
+    expect(actual).toEqual('SELECT DISTINCT "col1", "col2" FROM "table1";')
   })
 
-  it('Produces [SELECT ALL "col1", "col2" FROM "testTable";]', () => {
+  it('Produces [SELECT ALL "col1", "col2" FROM "table1";]', () => {
     const actual = sql
       .selectAll(col1, col2)
       .from(table)
       .getSQL()
 
-    expect(actual).toEqual('SELECT ALL "col1", "col2" FROM "testTable";')
+    expect(actual).toEqual('SELECT ALL "col1", "col2" FROM "table1";')
   })
 
-  it('Produces [SELECT DISTINCT "col1", "col2" FROM "testTable";] using select first param', () => {
+  it('Produces [SELECT DISTINCT "col1", "col2" FROM "table1";] using select first param', () => {
     const actual = sql
       .select(DISTINCT, col1, col2)
       .from(table)
       .getSQL()
 
-    expect(actual).toEqual('SELECT DISTINCT "col1", "col2" FROM "testTable";')
+    expect(actual).toEqual('SELECT DISTINCT "col1", "col2" FROM "table1";')
   })
 
-  it('Produces [SELECT ALL "col1", "col2" FROM "testTable";] using select first param', () => {
+  it('Produces [SELECT ALL "col1", "col2" FROM "table1";] using select first param', () => {
     const actual = sql
       .select(ALL, col1, col2)
       .from(table)
       .getSQL()
 
-    expect(actual).toEqual('SELECT ALL "col1", "col2" FROM "testTable";')
+    expect(actual).toEqual('SELECT ALL "col1", "col2" FROM "table1";')
   })
 
-  it('Produces [SELECT ALL FROM "testTable";] only ALL is valid (as param)', () => {
+  it('Produces [SELECT ALL FROM "table1";] only ALL is valid (as param)', () => {
     const actual = sql
       .select(ALL)
       .from(table)
       .getSQL()
 
-    expect(actual).toEqual('SELECT ALL FROM "testTable";')
+    expect(actual).toEqual('SELECT ALL FROM "table1";')
   })
 
-  it('Produces [SELECT ALL FROM "testTable";] only ALL is valid (using selectAll())', () => {
+  it('Produces [SELECT ALL FROM "table1";] only ALL is valid (using selectAll())', () => {
     const actual = sql
       .selectAll()
       .from(table)
       .getSQL()
 
-    expect(actual).toEqual('SELECT ALL FROM "testTable";')
+    expect(actual).toEqual('SELECT ALL FROM "table1";')
   })
 })