+++ CHANGELOG.md
@@ -1,6 +1,20 @@
 # Changelog
 <!-- https://keepachangelog.com/en/1.0.0/ -->
 
+## [0.11.0]  2022-04-11
+### Changes
+- Table data builder change to array of FromItemInfo
+- testTable in unit test renamed to table1
+- Remove set builderOption from OrderByItemInfo and SelectItemInfo, not needed anymore as we are passing the builderData to getStmt()
+- Update RailRoad documentation to include Joins steps
+
+### Added
+- From step can take more than one table
+- Builder option addAsBeforeTableAlias to Builder Option
+- AliasedTable class, table can be aliased
+- CrossJoin Step
+- Join, InnerJoin, LeftJoin, RightJoin, FullOuterJoin Steps
+
 ## [0.10.1]  2022-04-08
 ### Changes
 - Rename "column?" to "col?" in unit test
@@ -12,14 +26,14 @@
 - Move Operand class into Operand.ts file
 - Rename models to types and move it inside models folder
 - Rename file select.ts to SelectItemInfo.ts
-- Add IStatementGiver Interface
-- Add option addPublicSchemaName to Builder Option
-- Add Builder data to Table getStmt
-- Add option addTableName to Builder Option
-- Add Builder data to Column getStmt
 
 ### Added
 - Having step can contain aggregate function condition
+- IStatementGiver Interface
+- option addPublicSchemaName to Builder Option
+- Builder data to Table getStmt
+- option addTableName to Builder Option
+- Builder data to Column getStmt
 
 ## [0.10.0]  2022-03-26
 ### Changes